03-aiprotocol/02-a2a/01_a2a_fundamentals/hello_a2a/pyproject.toml
[project]
name = "hello-a2a"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.1",
    "uvicorn>=0.35.0",
]
